requirements.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
boto3==1.34.0
django-storages==1.14.2
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
reportlab==4.1.0
qrcode==7.4.2
requests==2.31.0
gunicorn==21.2.0
